Vue : Login / Loggout / Conditionnal navbar / conditionnal footer

+++ resources/js/application/layouts/layout.js
@@ -1,15 +1,35 @@
-import token from '../services/token.js';
+import { apiService } from '../_services/apiService.js';
 
 export default {
     data() {
         return {
-            isLogged: token.isTokenStored()
+            connected: localStorage.getItem('zotToken') != null ? true : false,
+            registerPath: '/inscription',
+            loginPath: '/connexion'
         }
     },
+
     methods: {
-        deconnexion() {
-            localStorage.clear();
-            location.href = '/login';
+        async deconnexion() {
+            try {
+                const graphqlQuery = {
+                    query: `mutation{logout}`,
+                };
+                const loggoutRequest = await apiService.post(`${location.origin}/graphql`, graphqlQuery);
+                const loggoutData = loggoutRequest.data.data.logout;
+                if (loggoutData) {
+                    this.connected = false;
+                    localStorage.clear();
+                    this.$router.push('/connexion')
+                }
+            } catch (error) {
+                this.connected = false;
+                localStorage.clear();
+                this.$router.push('/connexion')
+            }
+        },
+        updateNavbar(isLogged){
+            this.connected = isLogged;
         }
     }
 }
